+++ src/index.js
@@ -1,0 +1,3 @@
+import ColorPicker from './Picker.svelte';
+
+export default ColorPicker;